avg. velocity with 0.5
git-svn-id: https://servisto.de/svn/share/trunk/scrum@3205 15bc23fa-3125-0410-a1f3-bdc40dbf1900

+++ src/main/java/scrum/mda/ScrumModelApplication.java
@@ -351,7 +351,7 @@ public class ScrumModelApplication extends AGeneratorApplication {
 				"editable").setSearchable(true);
 			sprintModel.addProperty("begin", Date.class).setEditablePredicate("datesEditable");
 			sprintModel.addProperty("end", Date.class).setEditablePredicate("datesEditable");
-			sprintModel.addProperty("velocity", Integer.class);
+			sprintModel.addProperty("velocity", Float.class);
 			sprintModel.addStringProperty("completedRequirementLabels").setRichtext(true).setSearchable(true);
 			sprintModel.addStringProperty("planningNote").setRichtext(true).setTemplateAvailable(true)
 					.setEditablePredicate("planningEditable").setSearchable(true);
